+++ test/dockerValidator.test.ts
@@ -320,6 +320,17 @@ function assertHealthcheckTypeUnknown(diagnostic: Diagnostic, type: string, star
     assert.equal(diagnostic.range.end.character, endCharacter);
 }
 
+function assertADDDestinationNotDirectory(diagnostic: Diagnostic, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
+    assert.equal(diagnostic.code, ValidationCode.INVALID_DESTINATION);
+    assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
+    assert.equal(diagnostic.source, source);
+    assert.equal(diagnostic.message, Validator.getDiagnosticMessage_ADDDestinationNotDirectory());
+    assert.equal(diagnostic.range.start.line, startLine);
+    assert.equal(diagnostic.range.start.character, startCharacter);
+    assert.equal(diagnostic.range.end.line, endLine);
+    assert.equal(diagnostic.range.end.character, endCharacter);
+}
+
 function assertADDRequiresAtLeastTwoArguments(diagnostic: Diagnostic, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
     assert.equal(diagnostic.code, ValidationCode.ARGUMENT_REQUIRES_AT_LEAST_TWO);
     assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
@@ -1517,6 +1528,12 @@ describe("Docker Validator Tests", function() {
             it("ok", function() {
                 let diagnostics = validateDockerfile("FROM alpine\nADD . .");
                 assert.equal(diagnostics.length, 0);
+
+                diagnostics = validateDockerfile("FROM alpine\nADD Dockerfile Dockerfile2 /root/");
+                assert.equal(diagnostics.length, 0);
+
+                diagnostics = validateDockerfile("#escape=`\nFROM microsoft/nanoserver\nADD Dockerfile Dockerfile2 C:\\tmp\\");
+                assert.equal(diagnostics.length, 0);
             });
 
             it("requires at least two", function() {
@@ -1535,6 +1552,16 @@ describe("Docker Validator Tests", function() {
                 diagnostics = validateDockerfile("FROM alpine\nADD --chown=root:root .");
                 assert.equal(diagnostics.length, 1);
                 assertADDRequiresAtLeastTwoArguments(diagnostics[0], 1, 22, 1, 23);
+            });
+
+            it("destination not a directory", function() {
+                let diagnostics = validateDockerfile("FROM alpine\nADD Dockerfile Dockerfile2 /root");
+                assert.equal(diagnostics.length, 1);
+                assertADDDestinationNotDirectory(diagnostics[0], 1, 27, 1, 32);
+
+                diagnostics = validateDockerfile("#escape=`\nFROM microsoft/nanoserver\nADD Dockerfile Dockerfile2 C:\\tmp");
+                assert.equal(diagnostics.length, 1);
+                assertADDDestinationNotDirectory(diagnostics[0], 2, 27, 2, 33);
             });
         });